Update RELEASE-PLAN.md
Hi Martin - I admit that although I started by just changing a few typos, my desire to make this a page that is easy for us to maintain and update kicked in and I did make somewhat substantial formatting changes. The double tables within the framework section confused me and then I realized I could just copy the format you made for the Tech Spec releases, and it would make it much easier to maintain. I am really glad we have this page now, and will be sure to keep it updated as it will be a fantastic resource for the community!! Thanks again for putting it together :)